Add files via upload
This Android application was created in Android studio using the Java language. This application houses the neural network I created and prototyped in MATLAB which had to be reimplemented from first principles in Java. This application also showcases interface design using both the drag and drop functionality in Android Studio as well as creating one from scratch using the Canvas class, where the screen had to be drawn again for each frame.